move tequilapi-metrics out of libraries (Waldz requirement)
