+++ src/dependencies/mysterium-tequilapi.js
@@ -20,7 +20,7 @@ import axios from 'axios'
 import type {Container} from '../app/di'
 import {BugReporterMetrics} from '../app/bug-reporting/bug-reporter-metrics'
 import HttpTequilapiClient from '../libraries/mysterium-tequilapi/client'
-import TequilapiClientWithMetrics from '../libraries/tequilapi-metrics'
+import TequilapiClientWithMetrics from '../app/bug-reporting/tequilapi-metrics'
 import {TIMEOUT_DEFAULT} from '../libraries/mysterium-tequilapi/timeouts'
 import type {TequilapiClient} from '../libraries/mysterium-tequilapi/client'
 import AxiosAdapter from '../libraries/mysterium-tequilapi/adapters/axios-adapter'
