+++ test/unit/app/bug-reporting/tequilapi-metrics.spec.js
@@ -17,12 +17,12 @@
 
 // @flow
 
-import {expect, beforeEach, describe, it} from '../../helpers/dependencies'
-import TequilapiClientWithMetrics from '../../../src/libraries/tequilapi-metrics'
-import {BugReporterMetrics, METRICS} from '../../../src/app/bug-reporting/bug-reporter-metrics'
-import EmptyTequilapiClientMock from '../renderer/store/modules/empty-tequilapi-client-mock'
-import {MapSync} from '../../../src/libraries/map-sync'
-import type {Metric} from '../../../src/app/bug-reporting/bug-reporter-metrics'
+import {expect, beforeEach, describe, it} from '../../../helpers/dependencies'
+import TequilapiClientWithMetrics from '../../../../src/app/bug-reporting/tequilapi-metrics'
+import {BugReporterMetrics, METRICS} from '../../../../src/app/bug-reporting/bug-reporter-metrics'
+import EmptyTequilapiClientMock from '../../renderer/store/modules/empty-tequilapi-client-mock'
+import {MapSync} from '../../../../src/libraries/map-sync'
+import type {Metric} from '../../../../src/app/bug-reporting/bug-reporter-metrics'
 
 describe('HttpTequilapiClientWithMetrics', () => {
   let api
